% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_most_shared}
\alias{get_most_shared}
\title{Use Buzzsumo API to get the most shared content for a domain}
\usage{
get_most_shared(url, api_key)
}
\arguments{
\item{url}{Domain}

\item{api_key}{Buzzsumo API key}
}
\value{
A nested list of results
}
\description{
Use Buzzsumo API to get the most shared content for a domain
}
\examples{
\dontrun{
get_most_shared('slate.com', '123')
}
}

