% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_top_result}
\alias{parse_top_result}
\title{Create a dataframe containing the most shared article}
\usage{
parse_top_result(parsed)
}
\arguments{
\item{parsed}{A nested list; the result of running get_most_shared}
}
\value{
A dataframe with one row
}
\description{
Create a dataframe containing the most shared article
}
\examples{
\dontrun{
parsed <- get_most_shared('slate.com', '123')
parse_top_result(parsed)
}
}

